refactor: give EventsMapModal it's own file
