+++ src/components/EventsMap/components/index.ts
@@ -1,0 +1,1 @@
+export { EventsMapModal } from './EventsMapModal'
